Fix warnings on signed comparison

+++ src/p4est_vtk.c
@@ -648,7 +648,7 @@ p4est_vtk_write_point_datav (p4est_vtk_context_t * cont,
     SC_CHECK_ABORT (values[all]->elem_size == sizeof (double),
                     P4EST_STRING
                     "_vtk: Error: incorrect point scalar data type; scalar data must contain doubles.");
-    SC_CHECK_ABORT (values[all]->elem_count == cont->num_nodes,
+    SC_CHECK_ABORT (values[all]->elem_count == (size_t) cont->num_nodes,
                     P4EST_STRING
                     "_vtk: Error: incorrect point scalar data count; see "
                     P4EST_STRING ".h for more details.");
@@ -669,7 +669,7 @@ p4est_vtk_write_point_datav (p4est_vtk_context_t * cont,
     SC_CHECK_ABORT (values[all]->elem_size == sizeof (double),
                     P4EST_STRING
                     "_vtk: Error: incorrect point vector data type; vector data must contain doubles.");
-    SC_CHECK_ABORT (values[all]->elem_count == (3 * cont->num_nodes),
+    SC_CHECK_ABORT (values[all]->elem_count == 3 * (size_t) cont->num_nodes,
                     P4EST_STRING
                     "_vtk: Error: incorrect point vector data count; see "
                     P4EST_STRING ".h for more details.");
@@ -851,7 +851,7 @@ p4est_vtk_write_cell_datav (p4est_vtk_context_t * cont,
                     P4EST_STRING
                     "_vtk: Error: incorrect cell scalar data type; scalar data must contain doubles.");
     SC_CHECK_ABORT (values[all]->elem_count ==
-                    cont->p4est->local_num_quadrants,
+                    (size_t) cont->p4est->local_num_quadrants,
                     P4EST_STRING
                     "_vtk: Error: incorrect cell scalar data count; scalar data must contain exactly p4est->local_num_quadrants doubles.");
   }
@@ -872,7 +872,7 @@ p4est_vtk_write_cell_datav (p4est_vtk_context_t * cont,
                     P4EST_STRING
                     "_vtk: Error: incorrect cell vector data type; vector data must contain doubles.");
     SC_CHECK_ABORT (values[all]->elem_count ==
-                    3 * cont->p4est->local_num_quadrants,
+                    3 * (size_t) cont->p4est->local_num_quadrants,
                     P4EST_STRING
                     "_vtk: Error: incorrect cell vector data count; vector data must contain exactly 3*p4est->local_num_quadrants doubles.");
   }
